+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.19.0-SNAPSHOT
 
 * Added `workbook?` predicate.
+* Added `auto-size-column!` function to automatically adjust column width to contents.
 
 
 ## Version 1.18.0